pyproject.toml
[project]
name = "adk-ollama-samples"
version = "0.1.0"
description = "Ollama ADK Samples"
requires-python = ">=3.14"
readme = "README.md"

dependencies = [
    "arize-phoenix>=12.25.1",
    "chromadb>=0.4.21",
    "google-adk[a2a,eval]>=2.0.0a3",
    "httpx>=0.28.1",
    "ipywidgets>=8.0.0",
    "ipykernel>=6.31.0",
    "litellm>=1.82.6",
    "mcp[cli]>=1.13.0",
    "pypdf>=6.4.2",
    "sentence-transformers>=5.2.0",
    "langchain-text-splitters>=1.1.0",
    "openinference-instrumentation-google-adk>=0.1.10",
    "pytest-asyncio>=1.3.0",
    "pytest>=9.0.2",
    "ddgs>=9.0.0",
    "docker>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.11",
    "mypy>=1.20.1",
    "pre-commit>=3.7.0",
    "types-requests>=2.32.4.20260107",
]

# -----------------------------
# Ruff (linting + import sorting)
# -----------------------------
[tool.ruff]
line-length = 88
target-version = "py314"
fix = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # warnings
    "I",   # import sorting
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line length (handled by ruff format)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"


# -----------------------------
# Mypy (typing)
# -----------------------------
[tool.mypy]
python_version = "3.14"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
pretty = true